+++ packages/provider/src/index.ts
@@ -1,6 +1,5 @@
-export * from "./blockfrost/blockfrost.js";
-export * from "./kupmios/kupmios.js";
+export * from "./blockfrost.js";
+export * from "./kupmios.js";
 export * from "./emulator.js";
 export * from "./maestro.js";
-export * from "./koios/koios.js";
-export * as FetchProvider from "./fetch.js";
+export * from "./koios.js";