+++ packages/core-types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lucid-evolution/core-types
 
+## 0.1.15
+
+### Patch Changes
+
+- [#268](https://github.com/Anastasia-Labs/lucid-evolution/pull/268) [`f8f905a`](https://github.com/Anastasia-Labs/lucid-evolution/commit/f8f905aeda20b30a03ba69e4b062a80a7a5be7de) Thanks [@solidsnakedev](https://github.com/solidsnakedev)! - update providers structure and improve side effects
+
 ## 0.1.14
 
 ### Patch Changes