+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @lucid-evolution/provider
 
+## 0.1.47
+
+### Patch Changes
+
+- [#273](https://github.com/Anastasia-Labs/lucid-evolution/pull/273) [`f119e2f`](https://github.com/Anastasia-Labs/lucid-evolution/commit/f119e2f64c3d5869f9a2b44832f267c45a6567e0) Thanks [@hadelive](https://github.com/hadelive)! - fix sellect wallet from private key
+
+- Updated dependencies [[`f119e2f`](https://github.com/Anastasia-Labs/lucid-evolution/commit/f119e2f64c3d5869f9a2b44832f267c45a6567e0)]:
+  - @lucid-evolution/wallet@0.1.42
+
+## 0.1.46
+
+### Patch Changes
+
+- [#268](https://github.com/Anastasia-Labs/lucid-evolution/pull/268) [`f8f905a`](https://github.com/Anastasia-Labs/lucid-evolution/commit/f8f905aeda20b30a03ba69e4b062a80a7a5be7de) Thanks [@solidsnakedev](https://github.com/solidsnakedev)! - update providers structure and improve side effects
+
+- Updated dependencies [[`f8f905a`](https://github.com/Anastasia-Labs/lucid-evolution/commit/f8f905aeda20b30a03ba69e4b062a80a7a5be7de)]:
+  - @lucid-evolution/core-types@0.1.15
+  - @lucid-evolution/utils@0.1.37
+  - @lucid-evolution/wallet@0.1.41
+
 ## 0.1.45
 
 ### Patch Changes