+++ packages/wallet/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @lucid-evolution/wallet
 
+## 0.1.42
+
+### Patch Changes
+
+- [#273](https://github.com/Anastasia-Labs/lucid-evolution/pull/273) [`f119e2f`](https://github.com/Anastasia-Labs/lucid-evolution/commit/f119e2f64c3d5869f9a2b44832f267c45a6567e0) Thanks [@hadelive](https://github.com/hadelive)! - fix sellect wallet from private key
+
+## 0.1.41
+
+### Patch Changes
+
+- Updated dependencies [[`f8f905a`](https://github.com/Anastasia-Labs/lucid-evolution/commit/f8f905aeda20b30a03ba69e4b062a80a7a5be7de)]:
+  - @lucid-evolution/core-types@0.1.15
+  - @lucid-evolution/utils@0.1.37
+  - @lucid-evolution/sign_data@0.1.18
+
 ## 0.1.40
 
 ### Patch Changes